Creation of validators for: Bill, Egg, InventoryMovement, Order, Pay …

+++ src/main/java/com/goldeneggs/InventoryMovement/InventoryMovementServiceImpl.java
@@ -1,5 +1,7 @@
 package com.goldeneggs.InventoryMovement;
 
+import com.goldeneggs.Egg.Egg;
+import com.goldeneggs.Egg.EggValidator;
 import com.goldeneggs.Exception.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -15,6 +17,8 @@ public class InventoryMovementServiceImpl implements InventoryMovementService {
 
     @Autowired
     private InventoryMovementRepository inventoryMovementRepository;
+
+    private InventoryMovementValidator inventoryMovementValidator;
 
     /**
      * {@inheritDoc}
@@ -40,6 +44,7 @@ public class InventoryMovementServiceImpl implements InventoryMovementService {
      */
     @Override
     public InventoryMovement save(InventoryMovement inventoryMovement) {
+        validateInventoryMovementOrThrow(inventoryMovement);
         return inventoryMovementRepository.save(inventoryMovement);
     }
 
@@ -50,9 +55,18 @@ public class InventoryMovementServiceImpl implements InventoryMovementService {
      */
     @Override
     public InventoryMovement update(Long id, InventoryMovement updatedInventoryMovement) {
-        InventoryMovement existing = get(id); // Will throw exception if not foun
+        InventoryMovement existing = inventoryMovementRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Inventory movement with ID " + id + " not found."));
 
+        validateInventoryMovementOrThrow(updatedInventoryMovement);
+
+        existing.setEgg(updatedInventoryMovement.getEgg());
+        existing.setCombs(updatedInventoryMovement.getCombs());
+        existing.setMovementDate(updatedInventoryMovement.getMovementDate());
+        existing.setOrder(updatedInventoryMovement.getOrder());
+        existing.setUser(updatedInventoryMovement.getUser());
         return inventoryMovementRepository.save(existing);
+
     }
 
     /**
@@ -62,7 +76,24 @@ public class InventoryMovementServiceImpl implements InventoryMovementService {
      */
     @Override
     public void delete(Long id) {
-        InventoryMovement inventoryMovement = get(id); // Will throw exception if not found
+        if(!inventoryMovementRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Inventory movement with ID " + id + " not found.");
+        }
         inventoryMovementRepository.deleteById(id);
+    }
+
+    private void validateInventoryMovementOrThrow(InventoryMovement movement) {
+        if (!inventoryMovementValidator.validateEgg(movement.getEgg())) {
+            throw new IllegalArgumentException("Huevo no válido o no existente");
+        }
+        if (!InventoryMovementValidator.validateMovementDate(movement.getMovementDate())) {
+            throw new IllegalArgumentException("Fecha no válida");
+        }
+        if (!InventoryMovementValidator.validateCombs(movement.getCombs())) {
+            throw new IllegalArgumentException("Panales requeridos inválidos");
+        }
+        if (!inventoryMovementValidator.validateUser(movement.getUser())) {
+            throw new IllegalArgumentException("Usuario no válido");
+        }
     }
 }
